restructure the folder structure
